Increase shutdown timeout
1st attempt to address https://github.com/pytorch/pytorch/issues/59795.

The RPC calls in shutdown no longer able to finish within 5s. It could
be relevant to the additional checks and verifications we added for
RRef wrapped values. After increasing the timeout, the error disappeared
in my local dev env. I would like to test if this is sufficient for
CI test too.

ghstack-source-id: 2653aa0ecae44c4f8b7467089febc4c5b2edf537
Pull Request resolved: https://github.com/pytorch/pytorch/pull/59801