Add RRef support to ShardedTensor.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/59776

Overall design: https://github.com/pytorch/pytorch/issues/55207.

In this PR, I've added support to ShardedTensor such that it also creates RRefs
pointing to the remote shards if the RPC framework is initialized.

As a result, this provides more flexiblity for ShardedTensor such that users
can use collectives with local shards or use the RPC framework to interact with
remote shards.
ghstack-source-id: 131262010

Differential Revision: [D29020844](https://our.internmc.facebook.com/intern/diff/D29020844/)